Gabi

+++ it4kids/src/main/java/it4kids/domain/AbstractModel.java
@@ -1,8 +1,6 @@
 package it4kids.domain;
 
 /**
- * Abstract modelul asta e ca si al lui sebi , banuiesc ca si noi avem nevoie de
- * un ID si la quiz , si la question
  * 
  * @author Flaviu
  *
@@ -10,8 +8,8 @@ package it4kids.domain;
 public abstract class AbstractModel {
 	private long id;
 
-	public int getId() {
-		return (int) id;
+	public long getId() {
+		return id;
 	}
 
 	public void setId(long id) {
